robot-agent/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
httpx==0.27.0
PyYAML==6.0.1
tenacity==8.2.3
openai>=1.30.0
python-dotenv==1.0.1
